+++ Polyhedron/demo/Polyhedron/Scene_triangulation_3_item.h
@@ -31,6 +31,7 @@ using namespace CGAL::Three;
   Q_OBJECT
 public:
   typedef CGAL::qglviewer::ManipulatedFrame ManipulatedFrame;
+  static const int number_of_bitset = 4; // also defined in "shader_c3t3.frag" and "shader_c3t3_edges.frag"
 
   Scene_triangulation_3_item(bool display_elements = true);
   Scene_triangulation_3_item(const T3 t3, bool display_elements = true);
@@ -43,6 +44,7 @@ public:
 
 
   void setColor(QColor c) Q_DECL_OVERRIDE;
+  void setUseSubdomainColors(bool use_subdomain_colors);
 
   void invalidateOpenGLBuffers() Q_DECL_OVERRIDE;
 
@@ -145,7 +147,9 @@ public:
 
   QColor get_histogram_color(const double v) const;
 
+  void resetVisibleSubdomain();
   void switchVisibleSubdomain(int);
+  bool isVisibleSubdomain(int) const;
 
   void itemAboutToBeDestroyed(Scene_item *) Q_DECL_OVERRIDE;